+++ src/main/java/com/nred/nuclearcraft/compat/jei/JeiProcessorCategory.java
@@ -13,6 +13,7 @@ import mezz.jei.api.gui.builder.IRecipeSlotBuilder;
 import mezz.jei.api.gui.drawable.IDrawable;
 import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
 import mezz.jei.api.helpers.IGuiHelper;
+import mezz.jei.api.neoforge.NeoForgeTypes;
 import mezz.jei.api.recipe.IFocusGroup;
 import mezz.jei.api.recipe.RecipeType;
 import mezz.jei.api.recipe.category.IRecipeCategory;
@@ -27,6 +28,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
@@ -87,30 +89,30 @@ public class JeiProcessorCategory<T extends ProcessorRecipe> implements IRecipeC
     public void setRecipe(IRecipeLayoutBuilder builder, T recipe, IFocusGroup focuses) {
         for (int i = 0; i < recipe.itemIngredients.size(); i++) {
             ScreenPosition position = recipeViewerInfo.item_inputs().get(i);
-            builder.addInputSlot(position.x() + 1, position.y() + 1).addIngredients(recipe.itemIngredients.get(i).ingredient());
+            builder.addInputSlot(position.x() + 1, position.y() + 1).addItemStacks(Arrays.stream(recipe.itemIngredients.get(i).getItemsRaw()).toList());
         }
 
         for (int i = 0; i < recipe.fluidIngredients.size(); i++) {
             ScreenPosition position = recipeViewerInfo.fluid_inputs().get(i);
-            builder.addInputSlot(position.x() + 1, position.y() + 1).addFluidStack(recipe.getFluidIngredients().get(i).getFluids()[0].getFluid(), recipe.getFluidIngredients().get(i).amount()).setFluidRenderer(1, false, 16, 16);
+            builder.addInputSlot(position.x() + 1, position.y() + 1).addIngredients(NeoForgeTypes.FLUID_STACK, Arrays.stream(recipe.fluidIngredients.get(i).getFluidsRaw()).toList()).setFluidRenderer(1, false, 16, 16);
         }
 
         for (int i = 0; i < recipe.itemProducts.size(); i++) {
             ScreenPosition position = recipeViewerInfo.item_outputs().get(i);
-            IRecipeSlotBuilder temp = builder.addOutputSlot(position.x() + 1, position.y() + 1).addIngredients(recipe.itemProducts.get(i).ingredient());
             SizedChanceItemIngredient ingredient = recipe.getItemProducts().get(i);
-            temp.setSlotName("" + i).addRichTooltipCallback((recipeSlotView, tooltip) -> tooltip.add(Component.translatable(MODID + ".recipe_viewer.chance_output", ingredient.minStackSize(), ingredient.count(), NCMath.decimalPlaces(ingredient.getMeanStackSize(), 2))));
+            IRecipeSlotBuilder temp = builder.addOutputSlot(position.x() + 1, position.y() + 1).addItemStacks(Arrays.stream(ingredient.getItemsRaw()).toList());
+            if (ingredient.chancePercent() < 100) {
+                temp.setSlotName("" + i).addRichTooltipCallback((recipeSlotView, tooltip) -> tooltip.add(Component.translatable(MODID + ".recipe_viewer.chance_output", ingredient.minStackSize(), ingredient.count(), NCMath.decimalPlaces(ingredient.getMeanStackSize(), 2))));
+            }
         }
 
         int size = recipeViewerInfo.fluid_outputs().size() < 3 ? 24 : 16;
         for (int i = 0; i < recipe.fluidProducts.size(); i++) {
             ScreenPosition position = recipeViewerInfo.fluid_outputs().get(i);
             SizedChanceFluidIngredient ingredient = recipe.getFluidProducts().get(i);
+            IRecipeSlotBuilder temp = builder.addOutputSlot(position.x() + 1, position.y() + 1).addIngredients(NeoForgeTypes.FLUID_STACK, Arrays.asList(ingredient.getFluidsRaw())).setFluidRenderer(1, false, size, size);
             if (ingredient.chancePercent() < 100) {
-                builder.addOutputSlot(position.x() + 1, position.y() + 1).addFluidStack(ingredient.getFluids()[0].getFluid(), ingredient.amount()).setFluidRenderer(1, false, size, size)
-                        .setSlotName("" + i).addRichTooltipCallback((recipeSlotView, tooltip) -> tooltip.add(Component.translatable(MODID + ".recipe_viewer.chance_output", ingredient.minStackSize(), ingredient.amount(), NCMath.decimalPlaces(ingredient.getMeanStackSize(), 2))));
-            } else {
-                builder.addOutputSlot(position.x() + 1, position.y() + 1).addFluidStack(ingredient.getFluids()[0].getFluid(), ingredient.amount()).setFluidRenderer(1, false, size, size);
+                temp.setSlotName("" + i).addRichTooltipCallback((recipeSlotView, tooltip) -> tooltip.add(Component.translatable(MODID + ".recipe_viewer.chance_output", ingredient.minStackSize(), ingredient.amount(), NCMath.decimalPlaces(ingredient.getMeanStackSize(), 2))));
             }
         }
     }