+++ src/main/java/com/nred/nuclearcraft/recipe/RecipeStats.java
@@ -2,8 +2,8 @@ package com.nred.nuclearcraft.recipe;
 
 import com.nred.nuclearcraft.block_entity.processor.info.ProcessorMenuInfo;
 import com.nred.nuclearcraft.datamap.FissionModeratorData;
-import com.nred.nuclearcraft.handler.SizedChanceFluidIngredient;
 import com.nred.nuclearcraft.handler.BlockEntityMenuInfo;
+import com.nred.nuclearcraft.handler.SizedChanceFluidIngredient;
 import com.nred.nuclearcraft.handler.TileInfoHandler;
 import com.nred.nuclearcraft.recipe.NCRecipes.BasicProcessorRecipeHandler;
 import com.nred.nuclearcraft.util.NCMath;
@@ -96,7 +96,7 @@ public class RecipeStats {
 
     private static void setFissionMaxModeratorLineFlux() {
         fission_max_moderator_line_flux = 0;
-        for (FissionModeratorData moderatorData : BuiltInRegistries.ITEM.getDataMap(FISSION_MODERATOR_DATA).values()) {
+        for (FissionModeratorData moderatorData : BuiltInRegistries.BLOCK.getDataMap(FISSION_MODERATOR_DATA).values()) {
             if (moderatorData != null) {
                 fission_max_moderator_line_flux = Math.max(fission_max_moderator_line_flux, moderatorData.fluxFactor());
             }